+++ library/core/src/main/java/com/google/android/exoplayer2/extractor/mkv/MatroskaExtractor.java
@@ -256,14 +256,6 @@ public class MatroskaExtractor implements Extractor {
    */
   private static final int SUBRIP_PREFIX_END_TIMECODE_OFFSET = 19;
   /**
-   * A special end timecode indicating that a subrip subtitle should be displayed until the next
-   * subtitle, or until the end of the media in the case of the last subtitle.
-   * <p>
-   * Equivalent to the UTF-8 string: "            ".
-   */
-  private static final byte[] SUBRIP_TIMECODE_EMPTY =
-      new byte[] {32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32};
-  /**
    * The value by which to divide a time in microseconds to convert it to the unit of the last value
    * in a subrip timecode (milliseconds).
    */
@@ -303,14 +295,6 @@ public class MatroskaExtractor implements Extractor {
    * in an SSA timecode (1/100ths of a second).
    */
   private static final long SSA_TIMECODE_LAST_VALUE_SCALING_FACTOR = 10000;
-  /**
-   * A special end timecode indicating that an SSA subtitle should be displayed until the next
-   * subtitle, or until the end of the media in the case of the last subtitle.
-   * <p>
-   * Equivalent to the UTF-8 string: "          ".
-   */
-  private static final byte[] SSA_TIMECODE_EMPTY =
-      new byte[] {32, 32, 32, 32, 32, 32, 32, 32, 32, 32};
   /**
    * The format of an SSA timecode.
    */
@@ -1238,21 +1222,18 @@ public class MatroskaExtractor implements Extractor {
     if (track.trueHdSampleRechunker != null) {
       track.trueHdSampleRechunker.sampleMetadata(track, timeUs);
     } else {
-      if (CODEC_ID_SUBRIP.equals(track.codecId)) {
-        commitSubtitleSample(
-            track,
-            SUBRIP_TIMECODE_FORMAT,
-            SUBRIP_PREFIX_END_TIMECODE_OFFSET,
-            SUBRIP_TIMECODE_LAST_VALUE_SCALING_FACTOR,
-            SUBRIP_TIMECODE_EMPTY);
-      } else if (CODEC_ID_ASS.equals(track.codecId)) {
-        commitSubtitleSample(
-            track,
-            SSA_TIMECODE_FORMAT,
-            SSA_PREFIX_END_TIMECODE_OFFSET,
-            SSA_TIMECODE_LAST_VALUE_SCALING_FACTOR,
-            SSA_TIMECODE_EMPTY);
+      if (CODEC_ID_SUBRIP.equals(track.codecId) || CODEC_ID_ASS.equals(track.codecId)) {
+        if (durationUs == C.TIME_UNSET) {
+          Log.w(TAG, "Skipping subtitle sample with no duration.");
+        } else {
+          setSubtitleEndTime(track.codecId, durationUs, subtitleSample.data);
+          // Note: If we ever want to support DRM protected subtitles then we'll need to output the
+          // appropriate encryption data here.
+          track.output.sampleData(subtitleSample, subtitleSample.limit());
+          sampleBytesWritten += subtitleSample.limit();
+        }
       }
+
       if ((blockFlags & C.BUFFER_FLAG_HAS_SUPPLEMENTAL_DATA) != 0) {
         // Append supplemental data.
         int size = blockAddData.limit();
@@ -1480,51 +1461,58 @@ public class MatroskaExtractor implements Extractor {
     // the correct end timecode, which we might not have yet.
   }
 
-  private void commitSubtitleSample(Track track, String timecodeFormat, int endTimecodeOffset,
-      long lastTimecodeValueScalingFactor, byte[] emptyTimecode) {
-    setSubtitleSampleDuration(
-        subtitleSample.data,
-        blockDurationUs,
-        timecodeFormat,
-        endTimecodeOffset,
-        lastTimecodeValueScalingFactor,
-        emptyTimecode);
-    // Note: If we ever want to support DRM protected subtitles then we'll need to output the
-    // appropriate encryption data here.
-    track.output.sampleData(subtitleSample, subtitleSample.limit());
-    sampleBytesWritten += subtitleSample.limit();
-  }
-
   /**
-   * Formats {@code durationUs} using {@code timecodeFormat}, and sets it as the end timecode in
-   * {@code subtitleSampleData}.
+   * Overwrites the end timecode in {@code subtitleData} with the correctly formatted time derived
+   * from {@code durationUs}.
    *
    * <p>See documentation on {@link #SSA_DIALOGUE_FORMAT} and {@link #SUBRIP_PREFIX} for why we use
    * the duration as the end timecode.
+   *
+   * @param codecId The subtitle codec; must be {@link #CODEC_ID_SUBRIP} or {@link #CODEC_ID_ASS}.
+   * @param durationUs The duration of the sample, in microseconds.
+   * @param subtitleData The subtitle sample in which to overwrite the end timecode (output
+   *     parameter).
    */
-  private static void setSubtitleSampleDuration(
-      byte[] subtitleSampleData,
-      long durationUs,
-      String timecodeFormat,
-      int endTimecodeOffset,
-      long lastTimecodeValueScalingFactor,
-      byte[] emptyTimecode) {
+  private static void setSubtitleEndTime(String codecId, long durationUs, byte[] subtitleData) {
+    byte[] endTimecode;
+    int endTimecodeOffset;
+    switch (codecId) {
+      case CODEC_ID_SUBRIP:
+        endTimecode =
+            formatSubtitleTimecode(
+                durationUs, SUBRIP_TIMECODE_FORMAT, SUBRIP_TIMECODE_LAST_VALUE_SCALING_FACTOR);
+        endTimecodeOffset = SUBRIP_PREFIX_END_TIMECODE_OFFSET;
+        break;
+      case CODEC_ID_ASS:
+        endTimecode =
+            formatSubtitleTimecode(
+                durationUs, SSA_TIMECODE_FORMAT, SSA_TIMECODE_LAST_VALUE_SCALING_FACTOR);
+        endTimecodeOffset = SSA_PREFIX_END_TIMECODE_OFFSET;
+        break;
+      default:
+        throw new IllegalArgumentException();
+    }
+    System.arraycopy(endTimecode, 0, subtitleData, endTimecodeOffset, endTimecode.length);
+  }
+
+  /**
+   * Formats {@code timeUs} using {@code timecodeFormat}, and sets it as the end timecode in {@code
+   * subtitleSampleData}.
+   */
+  private static byte[] formatSubtitleTimecode(
+      long timeUs, String timecodeFormat, long lastTimecodeValueScalingFactor) {
+    Assertions.checkArgument(timeUs != C.TIME_UNSET);
     byte[] timeCodeData;
-    if (durationUs == C.TIME_UNSET) {
-      timeCodeData = emptyTimecode;
-    } else {
-      int hours = (int) (durationUs / (3600 * C.MICROS_PER_SECOND));
-      durationUs -= (hours * 3600 * C.MICROS_PER_SECOND);
-      int minutes = (int) (durationUs / (60 * C.MICROS_PER_SECOND));
-      durationUs -= (minutes * 60 * C.MICROS_PER_SECOND);
-      int seconds = (int) (durationUs / C.MICROS_PER_SECOND);
-      durationUs -= (seconds * C.MICROS_PER_SECOND);
-      int lastValue = (int) (durationUs / lastTimecodeValueScalingFactor);
+    int hours = (int) (timeUs / (3600 * C.MICROS_PER_SECOND));
+    timeUs -= (hours * 3600 * C.MICROS_PER_SECOND);
+    int minutes = (int) (timeUs / (60 * C.MICROS_PER_SECOND));
+    timeUs -= (minutes * 60 * C.MICROS_PER_SECOND);
+    int seconds = (int) (timeUs / C.MICROS_PER_SECOND);
+    timeUs -= (seconds * C.MICROS_PER_SECOND);
+    int lastValue = (int) (timeUs / lastTimecodeValueScalingFactor);
       timeCodeData = Util.getUtf8Bytes(String.format(Locale.US, timecodeFormat, hours, minutes,
           seconds, lastValue));
-    }
-    Assertions.checkState(timeCodeData.length == emptyTimecode.length);
-    System.arraycopy(timeCodeData, 0, subtitleSampleData, endTimecodeOffset, timeCodeData.length);
+    return timeCodeData;
   }
 
   /**